Lobby Layout Prototype
ADDED:
Cases on the right.
Plants create pathway to button.
Office area on left.

TO DO:
Cases on right when you enter need to be filled.
Room needs to be shrunk in response to average player size.
Need to add small detail to the left half of room (ex: loose paper).